Classify Tescia diagnostic modules
