+++ src/agilab/apps/builtin/tescia_diagnostic_project/src/tescia_diagnostic/compat/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility helpers for classified TeSciA diagnostic modules."""
