+++ src/agilab/apps/builtin/tescia_diagnostic_project/src/tescia_diagnostic/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Classified TeSciA diagnostic package modules."""
